change the function createCabin to createEditCabin so it can perform both actions

+++ src/services/apiCabins.js
@@ -28,34 +28,44 @@ export async function deleteCabin(id) {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  // url image example
-  // https://xvcjrmdztxtrrudzavok.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
+// url image example
+// https://xvcjrmdztxtrrudzavok.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
+
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
 
-  // create a cabin
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
-    console.log(error)
-    throw new Error('Cabin could not be created')
+    console.error(error);
+    throw new Error("Cabin could not be created");
   }
 
-  // upload an image
-  const { error: storageError } = await supabase
-    .storage
-    .from('cabin-images')
-    .upload(imageName, newCabin.image)
+  // 2. Upload image
+  if (hasImagePath) return data;
 
-  // Delete the cabin IF there was an error uplaoding image
+  const { error: storageError } = await supabase.storage
+    .from("cabin-images")
+    .upload(imageName, newCabin.image);
+
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
